Task #3942: Merged branch into the trunk. Fixed one more warning in OnlineControl.cc that was introduced on the trunk after the task branch had been created.

+++ MAC/APL/PAC/ITRFBeamServer/src/BeamServer.cc
@@ -1271,7 +1271,7 @@ DigitalBeam* BeamServer::checkBeam(GCFPortInterface* 				port,
 	}
 
 	// nr of subbands should fit in the beamlet space.
-	if (allocation.getSubbandBitset().count() > itsCurrentMaxBeamlets) {
+	if (static_cast<int>(allocation.getSubbandBitset().count()) > itsCurrentMaxBeamlets) {
 		LOG_ERROR_STR("Too many subbands specified (" << allocation.getSubbandBitset().count() << ") only " 
 					<< itsCurrentMaxBeamlets << " allowed");
 		return (0);
